add users (#6)

+++ app/jobs/create_story_pages_job.rb
@@ -3,18 +3,26 @@
 class CreateStoryPagesJob < ApplicationJob
   queue_as :default
 
-  def perform(story_id)
+  def perform(story_id:, current_user_id:)
+    user = User.find(current_user_id)
+    return if user.open_ai_token.blank?
+
+    @client = OpenAI::Client.new(access_token: user.open_ai_token)
     story = Story.find(story_id)
     story.pages.destroy_all if story.pages.any?
 
     create_pages(story)
     story.broadcast_pages
-    CreateStoryExcerptJob.perform_later(story_id)
-    CreatePageImagesJob.perform_later(story_id, story.pages.first.id)
-    CreatePageImagesJob.perform_later(story_id, story.pages.second.id)
+    enqueue_subsequent_jobs(story:, story_id:, current_user_id:)
   end
 
   private
+
+  def enqueue_subsequent_jobs(story:, story_id:, current_user_id:)
+    CreateStoryExcerptJob.perform_later(story_id:, current_user_id:)
+    CreatePageImagesJob.perform_later(story_id:, page_id: story.pages.first.id, current_user_id:)
+    CreatePageImagesJob.perform_later(story_id:, page_id: story.pages.second.id, current_user_id:)
+  end
 
   def create_pages(story)
     response = create_pages_content(story)
@@ -29,9 +37,7 @@ class CreateStoryPagesJob < ApplicationJob
   end
 
   def create_pages_content(story)
-    client = OpenAI::Client.new
-
-    client.completions(
+    @client.completions(
       parameters: {
         model: 'text-davinci-003',
         prompt: story.prompt,
